Update printf formatting
- MSVC does not yet implement z and t sizes.
- Modify new-style code to use std::cout.
- Preserved fprintf for old-style code, and casted size_t/ptrdiff_t
variables to 64-bit.

+++ SurgSim/Testing/MlcpIO/ReadText.cpp
@@ -267,8 +267,8 @@ static bool readEigenMatrix(const std::string& fileName, FILE* in, const char* l
 			ptrdiff_t newNumRows = matrix->rows() + 1;
 			if (rowVector.cols() != numCols)
 			{
-				fprintf(stderr, "Inconsistent number of columns for Eigen matrix (%td vs %td)\n  in file '%s'\n",
-						numCols, rowVector.cols(), fileName.c_str());
+				fprintf(stderr, "Inconsistent number of columns for Eigen matrix (%lld vs %lld)\n  in file '%s'\n",
+						static_cast<int64_t>(numCols), static_cast<int64_t>(rowVector.cols()), fileName.c_str());
 				return false;
 			}
 			matrix->conservativeResize(newNumRows, numCols);
